+++ playwright.config.ts
@@ -1,6 +1,11 @@
 import { defineConfig } from '@playwright/test';
 
 export default defineConfig({
-	webServer: { command: 'npm run build && npm run preview', port: 4173 },
+	webServer: {
+		command: 'npm run build && npm run preview',
+		port: 4173,
+		timeout: 360000,
+		reuseExistingServer: !process.env.CI
+	},
 	testMatch: '**/*.e2e.{ts,js}'
 });